modal is displaying
