+++ src/javascripts/helpers/data/boardData.js
@@ -19,5 +19,7 @@ const getBoards = () => new Promise((resolve, reject) => {
 
 const deleteBoard = (boardId) => axios.delete(`${baseUrl}/boards/${boardId}.json`);
 
+const addBoard = (newBoard) => axios.post(`${baseUrl}/boards.json`, newBoard);
 
-export default { getBoards, deleteBoard };
+
+export default { getBoards, deleteBoard, addBoard };
