+++ src/javascripts/main.js
@@ -3,6 +3,7 @@ import apiKeys from './helpers/apiKeys.json';
 import authData from './helpers/data/authData';
 import auth from './components/auth/auth';
 import navbar from './components/navbar/navbar';
+import 'bootstrap';
 import '../styles/main.scss';
 
 const init = () => {
